RapidXML & ArenaObjects
A lot of stuff.

RapidXML is an XML parser that is really fast (like the name suggests, and is even used by phones).

Arena is a thing that loads all of the data of an area from an XML document.
The benefit of this is ability to make custom maps.

So far I just have it to where it can load walls and music. That's it.