Save Data & Filters (Organization)
I'm excited. This game is starting to take hold. Now (if you ignore the hitbox error that makes all of this irrelavent and not possible), you can walk around blindly on an empty screen based off an untested XML area parser and then save the position in a non-tested manor!

I also added Visual Studio Filters to the files to make it easier to look for them. Things are far more organized now.